New Generic Tree
