+++ uberfire-extensions/uberfire-widgets/uberfire-widgets-core/uberfire-widgets-core-client/src/main/java/org/uberfire/ext/widgets/core/client/tree/GenericTree.java
@@ -1,0 +1,149 @@
+/*
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.uberfire.ext.widgets.core.client.tree;
+
+import java.util.Iterator;
+
+import com.google.gwt.event.logical.shared.CloseEvent;
+import com.google.gwt.event.logical.shared.CloseHandler;
+import com.google.gwt.event.logical.shared.HasCloseHandlers;
+import com.google.gwt.event.logical.shared.HasOpenHandlers;
+import com.google.gwt.event.logical.shared.HasSelectionHandlers;
+import com.google.gwt.event.logical.shared.OpenEvent;
+import com.google.gwt.event.logical.shared.OpenHandler;
+import com.google.gwt.event.logical.shared.SelectionEvent;
+import com.google.gwt.event.logical.shared.SelectionHandler;
+import com.google.gwt.event.shared.HandlerRegistration;
+import com.google.gwt.user.client.ui.Composite;
+import com.google.gwt.user.client.ui.FlowPanel;
+import com.google.gwt.user.client.ui.IsWidget;
+import org.uberfire.ext.widgets.core.client.resources.TreeNavigatorResources;
+
+public class GenericTree<T extends GenericTreeItem> extends Composite implements HasSelectionHandlers<T>,
+                                                      HasOpenHandlers<T>,
+                                                      HasCloseHandlers<T> {
+
+    final FlowPanel container = new FlowPanel();
+
+    private T  curSelection = null;
+
+    public GenericTree() {
+        initWidget(container);
+        container.setStylePrimaryName(TreeNavigatorResources.INSTANCE.css().tree());
+    }
+
+    @Override
+    public HandlerRegistration addOpenHandler(final OpenHandler<T> handler) {
+        return addHandler(handler,
+                          OpenEvent.getType());
+    }
+
+    @Override
+    public HandlerRegistration addCloseHandler(final CloseHandler<T> handler) {
+        return addHandler(handler,
+                          CloseEvent.getType());
+    }
+
+    @Override
+    public HandlerRegistration addSelectionHandler(final SelectionHandler<T> handler) {
+        return addHandler(handler,
+                          SelectionEvent.getType());
+    }
+
+    public void clear() {
+        container.clear();
+    }
+
+    void fireStateChanged(final T item,
+                          final T.State state) {
+        if (state.equals(T.State.OPEN)) {
+            OpenEvent.fire(this,
+                           item);
+        } else {
+            CloseEvent.fire(this,
+                            item);
+        }
+        onSelection(item,
+                    true);
+    }
+
+    public void setSelectedItem(final T item,
+                                final boolean fireEvents) {
+        onSelection(item,
+                    fireEvents);
+    }
+
+    public GenericTreeItem getSelectedItem() {
+        return curSelection;
+    }
+
+    public void setSelectedItem(final T item) {
+        onSelection(item,
+                    true);
+    }
+
+    void onSelection(final T item,
+                     final boolean fireEvents) {
+        if (curSelection != null) {
+            curSelection.setSelected(false);
+        }
+        curSelection = item;
+
+        if (curSelection != null) {
+            // Select the item and fire the selection event.
+            curSelection.setSelected(true);
+            if (fireEvents) {
+                SelectionEvent.fire(this,
+                                    curSelection);
+            }
+        }
+    }
+
+    public GenericTreeItem addItem(final GenericTreeItem.Type type,
+                                   final String value,
+                                   final String label,
+                                   final IsWidget icon) {
+        final GenericTreeItem item = new GenericTreeItem(type,
+                                                         value,
+                                                         label,
+                                                         icon);
+        return addItem(item);
+    }
+
+    public GenericTreeItem addItem(final GenericTreeItem item) {
+        container.add(item);
+        item.setGenericTree(this);
+        return item;
+    }
+
+    public void removeItem(final GenericTreeItem genericTreeItem) {
+        container.remove(genericTreeItem);
+    }
+
+    public Iterable<GenericTreeItem> getGenericItems() {
+        return new Iterable<GenericTreeItem>() {
+            @Override
+            public Iterator<GenericTreeItem> iterator() {
+                return new GenericTreeItem.TreeItemIterator(container);
+            }
+        };
+    }
+
+    public boolean isEmpty() {
+        return container.getWidgetCount() == 0;
+    }
+}
